add LocalState
