+++ lib/src/redux_component/redux_component.dart
@@ -7,7 +7,7 @@ export 'dependent.dart';
 export 'dispatch_bus.dart';
 export 'helper.dart';
 export 'lifecycle.dart';
-export 'local_props.dart';
+export 'local.dart';
 export 'logic.dart';
 export 'page.dart';
 export 'private_reducer_mixin.dart';
